Use 64-bit timestamps after all
If SDL can't do it, we will.

+++ src/gui/Timestamp.h
@@ -5,5 +5,5 @@
 
 namespace gui
 {
-	using Timestamp = int32_t;
+	using Timestamp = int64_t;
 }
